Rename case table insert button to case table manage button
