+++ frontend/src/components/cases/case-table-manage-button.tsx
